gateway/requirements.txt
web3
asyncio
argparse
pyyaml
redis
requests
bip_utils
gql